Version 1 Done

+++ controllers/managementController.js
@@ -1,6 +1,10 @@
 const User = require('../models/User');
 
 exports.createUser = async (req, res) => {
+  // Check if user is logged in
+  if (!req.session.user) {
+    return res.redirect('/auth/login');
+  }
   const { username, password, userType, leadAccess } = req.body;
   try {
     const existingUser = await User.findOne({ username });
@@ -21,6 +25,10 @@ exports.createUser = async (req, res) => {
 };
 
 exports.getUsers = async (req, res) => {
+  // Check if user is logged in
+  if (!req.session.user) {
+    return res.redirect('/auth/login');
+  }
   try {
     if (req.session.user.userType === 'agent') {
       return res.status(403).send('Unauthorized');
@@ -37,18 +45,27 @@ exports.getUsers = async (req, res) => {
 };
 
 exports.editUser = async (req, res) => {
+  // Check if user is logged in
+  if (!req.session.user) {
+    return res.redirect('/auth/login');
+  }
   try {
     const user = await User.findById(req.params.userId);
     if (!user) {
       return res.status(404).send('User not found');
     }
-    res.render('editmanagement', { user: user });
+    const userType=req.session.user.userType
+    res.render('editmanagement', { user: user ,userType});
   } catch (err) {
     res.status(500).send(err.message);
   }
 };
 
 exports.updateUser = async (req, res) => {
+  // Check if user is logged in
+  if (!req.session.user) {
+    return res.redirect('/auth/login');
+  }
   try {
     const user = await User.findByIdAndUpdate(req.params.userId, req.body, { new: true });
     if (!user) {
@@ -61,6 +78,10 @@ exports.updateUser = async (req, res) => {
 };
 
 exports.deleteUser = async (req, res) => {
+  // Check if user is logged in
+  if (!req.session.user) {
+    return res.redirect('/auth/login');
+  }
   try {
     const user = await User.findByIdAndDelete(req.params.userId);
     if (!user) {
